cat: Fix trust increase after playing
`add_trust` was hard capped at 5. As both feeding and playing with the
cat used `add_trust` it was not possible to reach trust > 5. But for
following orders, trust > 5 is required.

We allow to pass the max achievable trust as parameter to `add_trust`,
and tune playing with the cat to allow reaching trust 10.
This allows (by feeding and afterwards playing with it) to reach a
trust > 5 for a cat. Then, the cat will follow orders.